requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
blis==0.7.11
brotli==1.1.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.2.2
cloudpathlib==0.16.0
confection==0.1.5
contourpy==1.3.2
cycler==0.12.1
cymem==2.0.11
distro==1.9.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
exceptiongroup==1.3.0
fastapi==0.116.1
ffmpy==0.6.1
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-vision==3.4.5
googleapis-common-protos==1.70.0
gradio==5.39.0
gradio-client==1.11.0
groovy==0.1.2
grpcio==1.74.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kiwisolver==1.4.8
langcodes==3.5.0
language-data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
murmurhash==1.0.13
numpy==1.24.3
openai==1.98.0
opencv-python==4.8.1.78
orjson==3.11.1
packaging==25.0
pandas==2.3.1
pillow==10.0.1
preshed==3.0.10
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pydub==0.25.1
pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
requests==2.32.4
rich==14.1.0
rsa==4.9.1
ruff==0.12.7
safehttpx==0.1.6
scikit-learn==1.7.1
scipy==1.11.4
seaborn==0.13.2
semantic-version==2.10.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.47.2
thinc==8.2.5
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.1.0
weasel==0.3.4
websockets==15.0.1
